LMW-153 Do not restore screenlet id when it was already generated
The screenletId is restored only if it was not generated yet. If the
screenletId already exists at this point, it means that an interactor
is using it, so we cannot restore the previous value. As a side
effect, any previous executing task will not deliver the result to
the new interactor. To avoid this behavior, only call screenlet
methods after onStart() activity/fragment callback. This ensures that
onRestoreInstanceState was already called.